+++ packages/vm/src/modules/in_memory_cache.rs
@@ -108,12 +108,12 @@ mod tests {
         let key_size = mem::size_of::<Checksum>();
         assert_eq!(key_size, 32);
 
-        // A Module consists of a Store (2 Arcs) and an Arc to the Engine.
+        // A Module consists of a Store (3 Arcs) and an Arc to the Engine.
         // This is 3 * 64bit of data, but we don't get any guarantee how the Rust structs
         // Module and Store are aligned (https://doc.rust-lang.org/reference/type-layout.html#the-default-representation).
         // So we get this value by trial and error. It can change over time and across platforms.
         let value_size = mem::size_of::<Module>();
-        assert_eq!(value_size, 48);
+        assert_eq!(value_size, 56);
 
         // Just in case we want to go that route
         let boxed_value_size = mem::size_of::<Box<Module>>();