+++ CHANGELOG.md
@@ -28,7 +28,8 @@ and this project adheres to
   testing as well as integration tests of smart contracts.
 - cosmwasm-vm: More accurate error messages for op codes related to bulk memory
   operations, reference types, SIMD and the Threads extension.
-  
+- cosmwasm-vm: Update `wasmer` to `2.0.0-rc2`
+
 ### Fixed
 
 - comswasm-vm: Whitelisted the `i64.extend32_s` operation.